+++ app/model/club.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, Integer, String, Date
+from sqlalchemy import Column, ForeignKey, Integer, String, Date, Boolean
 from sqlalchemy.orm import relationship
 
 from app.db.session import Base
@@ -16,6 +16,7 @@ class Club(Base):
     skill = Column(String(24), comment="실력")
     img = Column(String(256), comment="클럽 이미지 URL")
     color = Column(String(24), comment="유니폼 색")
+    deleted = Column(Boolean, default=True, comment="삭제 여부")
 
     join_club = relationship("JoinClub", back_populates="club")
 
@@ -26,6 +27,7 @@ class JoinClub(Base):
     seq = Column(Integer, primary_key=True, comment="시퀀스")
     clubs_seq = Column(Integer, ForeignKey("clubs.seq"))
     user_seq = Column(Integer, ForeignKey("users.seq"))
+    role = Column(String(10), comment="역할")
 
     club = relationship("Club", back_populates="join_club")
     user = relationship("User", back_populates="join_club")
