+++ app/rest_api/schema/club.py
@@ -14,15 +14,16 @@ class UpdateClubSchema(BaseModel):
     skill: str = Field(title="실력", default=None)
     img: str = Field(title="클럽 이미지 URL", default=None)
     color: str = Field(title="유니폼 색", default=None)
+    deleted: bool = Field(title="팀 삭제", default=None)
 
 
 class ClubSchema(BaseModel):
-    name: str = Field(title="이름", default=None)
+    name: str = Field(title="이름")
     register_date: date = Field(title="창단일", default=None)
-    location: str = Field(title="활동 장소", default=None)
-    age_group: str = Field(title="연령대", default=None)
-    membership_fee: int = Field(title="회비", default=None)
-    skill: str = Field(title="실력", default=None)
+    location: str = Field(title="활동 장소")
+    age_group: str = Field(title="연령대")
+    membership_fee: int = Field(title="회비")
+    skill: str = Field(title="실력")
     img: str = Field(title="클럽 이미지 URL", default=None)
     color: str = Field(title="유니폼 색", default=None)
